refactor(#697): relocate remaining controllers from widgets/ to services/
- Relocate 5 non-widget files from widgets/ to services/:
  registration_controller, emoji_autocomplete_controller,
  mention_autocomplete_controller, voice_send_handler,
  member_sheet_launcher (all are controllers/handlers, not widgets)
- Remove matrix import from room_permissions_host.dart:
  create RoomPermissionsSyncWatcher service for SyncUpdate handling,
  move JoinRules conversion to PowerLevelService.setJoinRules
- Replace EventTypes constants with string literals
- Update all import paths in lib/ and test/
- flutter analyze: 0 issues, flutter test: 2318 pass
