+++ lib/features/auth/screens/registration_screen.dart
@@ -6,8 +6,8 @@ import 'package:kohera/core/routing/route_names.dart';
 import 'package:kohera/core/services/app_config.dart';
 import 'package:kohera/core/services/client_manager.dart';
 import 'package:kohera/core/services/matrix_service.dart';
+import 'package:kohera/features/auth/services/registration_controller.dart';
 import 'package:kohera/features/auth/widgets/app_logo_header.dart';
-import 'package:kohera/features/auth/widgets/registration_controller.dart';
 import 'package:kohera/features/auth/widgets/registration_views.dart';
 import 'package:kohera/shared/widgets/loading_button_child.dart';
 import 'package:provider/provider.dart';
