+++ lib/features/chat/screens/chat_screen.dart
@@ -24,9 +24,11 @@ import 'package:kohera/features/chat/screens/thread_screen.dart';
 import 'package:kohera/features/chat/services/chat_message_actions.dart';
 import 'package:kohera/features/chat/services/chat_search_controller.dart';
 import 'package:kohera/features/chat/services/compose_state_controller.dart';
+import 'package:kohera/features/chat/services/emoji_autocomplete_controller.dart';
 import 'package:kohera/features/chat/services/file_send_handler.dart';
 import 'package:kohera/features/chat/services/gif_send_handler.dart';
 import 'package:kohera/features/chat/services/linkable_span_builder.dart';
+import 'package:kohera/features/chat/services/mention_autocomplete_controller.dart';
 import 'package:kohera/features/chat/services/message_display_resolver.dart';
 import 'package:kohera/features/chat/services/message_forwarder.dart';
 import 'package:kohera/features/chat/services/message_timeline_controller.dart';
@@ -44,11 +46,9 @@ import 'package:kohera/features/chat/widgets/chat_app_bar.dart';
 import 'package:kohera/features/chat/widgets/compose_bar_section.dart';
 import 'package:kohera/features/chat/widgets/delete_event_dialog.dart';
 import 'package:kohera/features/chat/widgets/desktop_drop_wrapper.dart';
-import 'package:kohera/features/chat/widgets/emoji_autocomplete_controller.dart';
 import 'package:kohera/features/chat/widgets/emoji_picker_sheet.dart';
 import 'package:kohera/features/chat/widgets/forward_message_dialog.dart';
 import 'package:kohera/features/chat/widgets/join_call_banner.dart';
-import 'package:kohera/features/chat/widgets/mention_autocomplete_controller.dart';
 import 'package:kohera/features/chat/widgets/message_action_sheet.dart';
 import 'package:kohera/features/chat/widgets/message_bubble_context_menu.dart';
 import 'package:kohera/features/chat/widgets/message_list_view.dart';
@@ -58,8 +58,8 @@ import 'package:kohera/features/chat/widgets/sticker_picker_overlay.dart';
 import 'package:kohera/features/chat/widgets/typing_indicator.dart';
 import 'package:kohera/features/home/screens/home_shell.dart';
 import 'package:kohera/features/rooms/models/kohera_room_member.dart';
+import 'package:kohera/features/rooms/services/member_sheet_launcher.dart';
 import 'package:kohera/features/rooms/services/room_summary_resolver.dart';
-import 'package:kohera/features/rooms/widgets/member_sheet_launcher.dart';
 import 'package:provider/provider.dart';
 
 class ChatScreen extends StatefulWidget {
