+++ lib/features/chat/services/emoji_autocomplete_controller.dart
