+++ lib/features/chat/services/mention_autocomplete_controller.dart
