+++ lib/features/chat/widgets/compose_bar.dart
@@ -9,14 +9,14 @@ import 'package:kohera/core/services/preferences_service.dart';
 import 'package:kohera/core/services/sticker_pack_service.dart';
 import 'package:kohera/core/utils/platform_info.dart';
 import 'package:kohera/features/chat/models/kohera_reply_preview.dart';
+import 'package:kohera/features/chat/services/emoji_autocomplete_controller.dart';
+import 'package:kohera/features/chat/services/mention_autocomplete_controller.dart';
 import 'package:kohera/features/chat/services/typing_controller.dart';
 import 'package:kohera/features/chat/services/voice_recording_controller.dart';
 import 'package:kohera/features/chat/widgets/attachment_preview_bar.dart';
 import 'package:kohera/features/chat/widgets/edit_preview_banner.dart';
-import 'package:kohera/features/chat/widgets/emoji_autocomplete_controller.dart';
 import 'package:kohera/features/chat/widgets/emoji_suggestion_list.dart';
 import 'package:kohera/features/chat/widgets/link_preview_card.dart';
-import 'package:kohera/features/chat/widgets/mention_autocomplete_controller.dart';
 import 'package:kohera/features/chat/widgets/mention_suggestion_overlay.dart';
 import 'package:kohera/features/chat/widgets/message_bubble_link_preview.dart';
 import 'package:kohera/features/chat/widgets/recording_indicator.dart';
