+++ lib/features/rooms/services/power_level_service.dart
@@ -1,3 +1,4 @@
+import 'package:kohera/features/rooms/models/kohera_room_permissions.dart';
 import 'package:matrix/matrix.dart';
 
 /// A partial update to a room's `m.room.power_levels` state event.
@@ -110,7 +111,7 @@ class PowerLevelService {
     if (value != null) content[key] = value;
   }
 
-  static void _mergeMap(
+ static void _mergeMap(
     Map<String, Object?> content,
     String key,
     Map<String, int>? patch,
@@ -121,4 +122,16 @@ class PowerLevelService {
     existing.addAll(patch.map(MapEntry.new));
     content[key] = existing;
   }
+
+  /// Sets join rules on a room from a [KoheraJoinRule] value.
+  static Future<void> setJoinRules(Room room, KoheraJoinRule rule) async {
+    await room.setJoinRules(_toSdkJoinRules(rule));
+  }
+
+  static JoinRules _toSdkJoinRules(KoheraJoinRule rule) => switch (rule) {
+        KoheraJoinRule.public => JoinRules.public,
+        KoheraJoinRule.invite => JoinRules.invite,
+        KoheraJoinRule.knock => JoinRules.knock,
+        KoheraJoinRule.restricted => JoinRules.restricted,
+      };
 }
