+++ lib/features/rooms/services/room_details_controller.dart
@@ -14,11 +14,11 @@ import 'package:kohera/features/rooms/models/kohera_room_permissions.dart';
 import 'package:kohera/features/rooms/models/kohera_room_summary.dart';
 import 'package:kohera/features/rooms/services/invite_user_dialog_params.dart';
 import 'package:kohera/features/rooms/services/join_access_controller.dart';
+import 'package:kohera/features/rooms/services/member_sheet_launcher.dart';
 import 'package:kohera/features/rooms/services/room_member_list_resolver.dart';
 import 'package:kohera/features/rooms/services/room_permissions_resolver.dart';
 import 'package:kohera/features/rooms/services/shared_media_loader.dart';
 import 'package:kohera/features/rooms/widgets/invite_user_dialog.dart';
-import 'package:kohera/features/rooms/widgets/member_sheet_launcher.dart';
 import 'package:kohera/features/rooms/widgets/shared_media_section.dart';
 import 'package:kohera/shared/services/avatar_resolver.dart';
 import 'package:matrix/matrix.dart';
